pyproject.toml
[tool.poetry]
name = "firefly-bot"
version = "0.1.0"
description = "A Telegram bot to create transactions in Firefly III"
authors = ["Valmik Jangla <mail@valmik.in>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.5.1"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
